+++ gsoap/custom/qtime.cpp
@@ -99,15 +99,14 @@ QTime *soap_in_xsd__time(struct soap *soap, char const *tag, QTime *a, char cons
   return a;
 }
 
-char const *soap_xsd__time2s(struct soap *soap, QTime a)
+const char * soap_xsd__time2s(struct soap *soap, QTime a)
 {
   //hh:mm:ss.zzzZ
   if (!a.isValid())
     a = QTime(0, 0, 0, 0);
   QString str = a.toString("HH:mm:ss.zzzZ");
   QByteArray ba = str.toLatin1();
-  const char *s = ba.constData();
-  soap_strcpy(soap->tmpbuf, qstrlen(ba)+1, s);
+  soap_strcpy(soap->tmpbuf, qstrlen(ba)+1, ba.constData());
   return soap->tmpbuf;
 }