+++ gsoap/plugin/wstapi.h
@@ -52,6 +52,7 @@ A commercial use license is available from Genivia, Inc., contact@genivia.com
 #define WSTAPI_H
 
 #include "wsaapi.h"     /* also includes stdsoap2.h, soapH.h (replace with soapcpp2-generated *H.h file) */
+#include "wsseapi.h"
 #include "smdevp.h"     /* digest algos */
 #include "threads.h"    /* mutex for sequence database */
 
@@ -65,9 +66,14 @@ extern "C" {
 #endif
 
 SOAP_FMAC1 int SOAP_FMAC2 soap_wst_request_saml_token(struct soap *soap, const char *endpoint, int soapver, const char *applyto, const char *username, const char *password, saml1__AssertionType **saml1, saml2__AssertionType **saml2);
+SOAP_FMAC1 int SOAP_FMAC2 soap_wst_request_psha1_token(struct soap *soap, const char *endpoint, int soapver, const char *applyto, const char *username, const char *password, char *psha1, size_t psha1len);
 
 #ifdef __cplusplus
 }
 #endif
+
+SOAP_FMAC5 int SOAP_FMAC6 soap_call___wst__RequestSecurityToken(struct soap *soap, const char *soap_endpoint, const char *soap_action, struct wst__RequestSecurityTokenType *wst__RequestSecurityToken, struct wst__RequestSecurityTokenResponseType *wst__RequestSecurityTokenResponse);
+
+SOAP_FMAC5 int SOAP_FMAC6 soap_call___wst__RequestSecurityTokenCollection(struct soap *soap, const char *soap_endpoint, const char *soap_action, struct wst__RequestSecurityTokenCollectionType *wst__RequestSecurityTokenCollection, struct wst__RequestSecurityTokenResponseCollectionType *wst__RequestSecurityTokenResponseCollection);
 
 #endif